requirements/base.txt
dj-database-url==0.5.0
psycopg2-binary==2.9.1
python-decouple==3.4
Django==3.2
djangorestframework==3.13.1
django-filter==21.1
redis==4.3.3
django-celery-beat==2.3.0
django-celery-results==2.3.1
